feat(argv): add --bypass-op-check cmd line arg
If someone attempts to award a delta to OP (the poster of the thread)
the bot will not accept their delta. This commit adds a command line
argument '--bypass-op-check' to allow deltas to be awarded to OP.
Mainly used for testing, this will be useful as it saves testers time,
allowing them to use a command line argument instead of changing the
code to skip over this test.